requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
gunicorn==24.0.0  # Добавлен явно (обязательно для хостинга)
itsdangerous==2.2.0
Jinja2==3.1.6
python-dotenv==1.1.0
Werkzeug==3.1.3
boto3
botocore